% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareLM.r
\name{ad_test}
\alias{ad_test}
\title{Anderson-Darling test for Normality}
\usage{
ad_test(x)
}
\arguments{
\item{x}{a numeric vector of data values, the number of which must be 
greater than 7. Missing values are allowed.}
}
\value{
A list with class “htest” containing the following components:
\describe{
  \item{statistic}{the value of the Anderson-Darling statistic}
  \item{p.value}{the p-value for the test}
  \item{method}{the character string “Anderson-Darling normality test”}
  \item{data.name}{a character string giving the name(s) of the data}
}
}
\description{
Anderson-Darling test for Normality
}
\details{
Performs the Anderson-Darling test for Normality. Missing values 
are allowed.

This specific function actually comes from \link[nortest]{ad.test}.
}
\examples{
data(BrendonSmall, package = "bootcamp")
model.1 = lm(Sodium ~ Calories, data = BrendonSmall)
model.2 = lm(Sodium ~ poly(Calories, 2), BrendonSmall)
ad_test(model.1$residuals)
ad_test(model.2$residuals)
}
\references{
authored by Juergen Gross in \link[nortest]{ad.test}
}
