% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_beta.R
\name{add_beta}
\alias{add_beta}
\title{Add standardized coefficients to lm output}
\usage{
add_beta(mod)
}
\arguments{
\item{mod}{output from an \code{lm} model}
}
\value{
a \code{summary.lm} object
}
\description{
Add standardized coefficients to lm summary
}
\details{
Using the output from an \code{lm} model, this functions adds the 
standardized coefficients to the output summary.

For a little more flexibility, consider the \code{lm.beta} package, which 
this function in based on.
}
\examples{
data(mtcars)
mm <- lm(mpg ~ ., data = mtcars)
summary(mm)
add_beta(mm)
}
