% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootcamp-package.r
\docType{package}
\name{bootcamp}
\alias{bootcamp}
\title{bootcamp}
\description{
Package supporting the bootcamp course
}
\details{
Functions, datasets, and tutorials, useful for the bootcamp course.
}
\keyword{package}
