% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareLM.r
\name{compareLM}
\alias{compareLM}
\title{Compare fit statistics for lm models}
\usage{
compareLM(fits, ...)
}
\arguments{
\item{fits}{A series of model object names, separated by commas.}

\item{...}{Other arguments passed to \code{list}.}
}
\value{
A list of two objects: The series of model calls, and a data 
        frame of statistics for each model.
}
\description{
Produces a table of fit statistics for multiple lm models.
}
\details{
Produces a table of fit statistics for multiple lm models: 
          AIC, AICc, BIC, p-value (for the F-test), R-squared, and adjusted R-squared.
          
          Smaller values for AIC, AICc, and BIC indicate a better balance
          of goodness-of-fit of the model and the complexity of the
          model. The goal is to find a model that adequately explains the 
          data without having too many terms.
          
          BIC tends to choose models with fewer parameters relative to AIC.
          
          In the table, \code{anderson.p} refers to the
          p-value for the Anderson-Darling test of 
          Normality of the residuals of the model. 
          The value of the test statistic itself is not reported, since that 
          isn't informative by itself. If this is wanted, run \link{ad_test} 
          directly.
          
          For comparisons with AIC, etc., to be valid, both models must
          have the same data, without transformations, use the same 
          dependent variable, and be fit with the same method.
          They do not need to be nested.
          
          The function will fail if a model formula is
          longer than 500 characters.
}
\examples{
### Compare among polynomial models
data(BrendonSmall, package = "bootcamp")
BrendonSmall$Calories = as.numeric(BrendonSmall$Calories)
model.1 = lm(Sodium ~ Calories, data = BrendonSmall)
model.2 = lm(Sodium ~ poly(Calories, 2), data = BrendonSmall)
model.3 = lm(Sodium ~ poly(Calories, 3), data = BrendonSmall)
model.4 = lm(Sodium ~ poly(Calories, 4), data = BrendonSmall)
compareLM(model.1, model.2, model.3, model.4)

}
\references{
\url{http://rcompanion.org/handbook/I_10.html}
            \url{http://rcompanion.org/rcompanion/e_05.html}
}
\author{
Salvatore Mangiafico, \email{mangiafico@njaes.rutgers.edu}
}
