% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_lalonde.R
\docType{data}
\name{lalonde}
\alias{lalonde}
\title{Dataset lalonde}
\format{
A data frame with 445 observations on the following 12 variables.
\describe{
  \item{age}{age in years}
  \item{educ}{years of schooling}
  \item{black}{indicator variable for blacks}
  \item{hisp}{indicator variable for Hispanics}
  \item{married}{indicator variable for martial status}
  \item{nodegr}{indicator variable for high school diploma}
  \item{re74}{real earnings in 1974}
  \item{re75}{real earnings in 1975}
  \item{re78}{real earnings in 1978}
  \item{u74}{indicator variable for earnings in 1974 being zero}
  \item{u75}{indicator variable for earnings in 1975 being zero}
  \item{treat}{an indicator variable for treatment status}
  
}
}
\source{
\code{lalonde} dataset from the \code{arm} package.
}
\description{
Dataset used by Dehejia and Wahba (1999) to evaluate propensity score matching.
}
\details{
This is a dataset that has been altered from the original 
dataset \code{movies} in the \code{jtools} package. The edits have 
been made for pedagogical purposes. 
Please consult \code{\link[jtools]{movies}} for the full dataset.
}
\keyword{datasets}
